--- a/packages/ifpack2/doc/UsersGuide/options.tex
+++ b/packages/ifpack2/doc/UsersGuide/options.tex
@@ -193,7 +193,7 @@ The following parameters are used in the block relaxation methods:
     {See~\ref{s:relaxation}. Currently has no effect. }
 \ccc{block relaxation: decouple dofs}
     {bool}
-    {false}
+    {true}
     {Whether to separate blocks according to the different degrees of
      freedom (PDEs) at each node. This assumes that dofs/node is constant
      throughout the matrix. Each block will have the same sparsity