--- a/packages/ifpack2/doc/UsersGuide/options.tex
+++ b/packages/ifpack2/doc/UsersGuide/options.tex
@@ -193,7 +193,7 @@ The following parameters are used in the block relaxation methods:
     {See~\ref{s:relaxation}. Currently has no effect. }
 \ccc{block relaxation: decouple dofs}
     {bool}
-    {true}
+    {false}
     {Whether to separate blocks according to the different degrees of
      freedom (PDEs) at each node. This assumes that dofs/node is constant
      throughout the matrix. Each block will have the same sparsity
@@ -201,7 +201,7 @@ The following parameters are used in the block relaxation methods:
      For example, when using a line partitioner this
      enables the use of the tridiagonal container even if the matrix's
      bandwidth is greater than 3.
-     Decoupling is the line smoothing behavior used in ML.}
+     Decoupling matches the behavior of line smoothing in ML.}
 \ccc{partitioner: type}
     {string}
     {``linear''}