--- a/packages/ifpack2/src/Ifpack2_BandedContainer_decl.hpp
+++ b/packages/ifpack2/src/Ifpack2_BandedContainer_decl.hpp
@@ -46,7 +46,6 @@
 /// \brief Ifpack2::BandedContainer class declaration
 
 #include "Ifpack2_Container.hpp"
-#include "Ifpack2_Details_MultiVectorLocalGatherScatter.hpp"
 #include "Ifpack2_Details_LapackSupportsScalar.hpp"
 #include "Tpetra_MultiVector.hpp"
 #include "Tpetra_Map.hpp"
@@ -69,8 +68,7 @@ namespace Ifpack2 {
 /// </ol>
 /// BandedContainer stores the diagonal blocks as Banded matrices, and
 /// solves them using either LAPACK (for the four Scalar types that it
-/// supports) or a custom LU factorization (for Scalar types not
-/// supported by LAPACK).
+/// supports)
 ///
 /// As with Ifpack2::Container, <tt>MatrixType</tt> must be a
 /// specialization of Tpetra::RowMatrix.  Using a Banded matrix for
@@ -101,14 +99,11 @@ namespace Ifpack2 {
 /// to modify the extract() method, so that it translates explicitly
 /// between local row and column indices, instead of just assuming
 /// that they are the same.
-template<class MatrixType,
-         class LocalScalarType,
-         bool supportsLapackScalar = ::Ifpack2::Details::LapackSupportsScalar<LocalScalarType>::value>
-class BandedContainer;
 
 template<class MatrixType, class LocalScalarType>
-class BandedContainer<MatrixType, LocalScalarType, true> :
-    public Container<MatrixType> {
+class BandedContainer
+: public ContainerImpl<MatrixType, LocalScalarType>
+{
   //! @name Internal typedefs (private)
   //@{
 private:
@@ -133,14 +128,16 @@ private:
   //! The Node type of the input (global) matrix.
   typedef typename Container<MatrixType>::node_type node_type;
 
+  typedef typename Container<MatrixType>::STS STS;
+
   typedef typename Container<MatrixType>::mv_type mv_type;
   typedef typename Container<MatrixType>::map_type map_type;
   typedef Tpetra::MultiVector<local_scalar_type, local_ordinal_type, global_ordinal_type, node_type> local_mv_type;
   typedef typename Container<MatrixType>::vector_type vector_type;
-  typedef typename Container<MatrixType>::partitioner_type partitioner_type;
   typedef typename Container<MatrixType>::import_type import_type;
 
   typedef typename Container<MatrixType>::HostView HostView;
+  typedef typename ContainerImpl<MatrixType, LocalScalarType>::HostSubview HostSubview;
   typedef typename local_mv_type::dual_view_type::t_host HostViewLocal;
 
   static_assert(std::is_same<MatrixType,
@@ -180,31 +177,21 @@ public:
   ///       with kl+ku superdiagonals, as we need the addtional storage for the LU decomposition.
   BandedContainer (const Teuchos::RCP<const row_matrix_type>& matrix,
                    const Teuchos::Array<Teuchos::Array<local_ordinal_type> >& partitions,
-                   const Teuchos::RCP<const import_type>& importer,
-                   int OverlapLevel,
-                   scalar_type DampingFactor);
+                   const Teuchos::RCP<const import_type>&,
+                   bool pointIndexed);
 
   BandedContainer (const Teuchos::RCP<const row_matrix_type>& matrix,
-                   const Teuchos::Array<local_ordinal_type>& localRows);
+                   const Teuchos::Array<local_ordinal_type>& localRows,
+                   bool pointIndexed);
 
   //! Destructor (declared virtual for memory safety of derived classes).
-  virtual ~BandedContainer ();
+  virtual ~BandedContainer () {}
 
   //@}
   //! \name Get and set methods
   //@{
 
-  //! Whether the container has been successfully initialized.
-  virtual bool isInitialized () const {
-    return IsInitialized_;
-  }
-
-  //! Whether the container has been successfully computed.
-  virtual bool isComputed () const {
-    return IsComputed_;
-  }
-
-  //! Set all necessary parameters.
+  //! Set all necessary parameters (super- and sub-diagonal bandwidth)
   virtual void setParameters (const Teuchos::ParameterList& List);
 
   //@}
@@ -217,28 +204,11 @@ public:
   //! Initialize and compute each block.
   virtual void compute ();
 
+  //! Compute the bandwidth (kl_, ku_) of each block from the input matrix,
+  //! unless the bandwidth has already been set through setParameters().
+  void computeBandwidth();
+
   void clearBlocks();
-
-  //! Compute <tt>Y := alpha * M^{-1} X + beta*Y</tt>.
-  virtual void
-  apply (HostView& X,
-         HostView& Y,
-         int blockIndex,
-         int stride,
-         Teuchos::ETransp mode = Teuchos::NO_TRANS,
-         scalar_type alpha = Teuchos::ScalarTraits<scalar_type>::one(),
-         scalar_type beta = Teuchos::ScalarTraits<scalar_type>::zero()) const;
-
-  //! Compute <tt>Y := alpha * diag(D) * M^{-1} (diag(D) * X) + beta*Y</tt>.
-  virtual void
-  weightedApply (HostView& X,
-                 HostView& Y,
-                 HostView& D,
-                 int blockIndex,
-                 int stride,
-                 Teuchos::ETransp mode = Teuchos::NO_TRANS,
-                 scalar_type alpha = Teuchos::ScalarTraits<scalar_type>::one(),
-                 scalar_type beta = Teuchos::ScalarTraits<scalar_type>::zero()) const;
 
   //@}
   //! \name Miscellaneous methods
@@ -271,9 +241,6 @@ private:
   //! Copy constructor: Declared but not implemented, to forbid copy construction.
   BandedContainer (const BandedContainer<MatrixType, LocalScalarType>& rhs);
 
-  //! Extract the submatrix identified by the local indices set by the constructor.
-  void extract ();
-
   /// \brief Factor the extracted submatrix.
   ///
   /// Call this after calling extract().
@@ -288,260 +255,29 @@ private:
   /// \param X [in] Subset permutation of the input X of apply().
   /// \param Y [in] Subset permutation of the input/output Y of apply().
   void
-  applyImpl (HostViewLocal& X,
-             HostViewLocal& Y,
+  solveBlock(HostSubview& X,
+             HostSubview& Y,
              int blockIndex,
-             int stride,
              Teuchos::ETransp mode,
              const local_scalar_type alpha,
-             const local_scalar_type beta) const;
+             const local_scalar_type beta) const override;
 
   //! The local diagonal block, which compute() extracts.
   std::vector<Teuchos::SerialBandDenseMatrix<int, local_scalar_type> > diagBlocks_;
 
-  //! Temporary X vector used in apply().
-  mutable std::vector<HostViewLocal> X_local;
-
-  //! Temporary Y vector used in apply().
-  mutable std::vector<HostViewLocal> Y_local;
-
   //! Permutation array from LAPACK (GETRF).
   Teuchos::Array<int> ipiv_;
-
-  //! If \c true, the container has been successfully initialized.
-  bool IsInitialized_;
-
-  //! If \c true, the container has been successfully computed.
-  bool IsComputed_;
 
   Teuchos::Array<local_ordinal_type> kl_; //< number of subdiagonals
   Teuchos::Array<local_ordinal_type> ku_; //< number of superdiagonals
 
   //! Scalar data for all blocks
-  local_scalar_type* scalars_;
+  Teuchos::Array<local_scalar_type> scalars_;
 
   //! Offsets in scalars_ array for all blocks
   Teuchos::Array<local_ordinal_type> scalarOffsets_;
 };
 
-template<class MatrixType, class LocalScalarType>
-class BandedContainer<MatrixType, LocalScalarType, false> :
-    public Container<MatrixType> {
-  //! @name Internal typedefs (private)
-  //@{
-private:
-  /// \brief The first template parameter of this class.
-  ///
-  /// This must be either a Tpetra::RowMatrix specialization or a
-  /// Tpetra::CrsMatrix specialization.  It may have entirely
-  /// different template parameters (e.g., \c scalar_type) than
-  /// <tt>InverseType</tt>.
-  typedef MatrixType matrix_type;
-  //! The second template parameter of this class.
-  typedef LocalScalarType local_scalar_type;
-  //! The internal representation of LocalScalarType in Kokkos::View
-  typedef typename Kokkos::Details::ArithTraits<local_scalar_type>::val_type local_impl_scalar_type;
-
-  //! The type of entries in the input (global) matrix.
-  typedef typename Container<MatrixType>::scalar_type scalar_type;
-  //! The type of local indices in the input (global) matrix.
-  typedef typename Container<MatrixType>::local_ordinal_type local_ordinal_type;
-  //! The type of global indices in the input (global) matrix.
-  typedef typename Container<MatrixType>::global_ordinal_type global_ordinal_type;
-  //! The Node type of the input (global) matrix.
-  typedef typename Container<MatrixType>::node_type node_type;
-
-  typedef typename Container<MatrixType>::mv_type mv_type;
-  typedef typename Container<MatrixType>::map_type map_type;
-  typedef Tpetra::MultiVector<local_scalar_type, local_ordinal_type, global_ordinal_type, node_type> local_mv_type;
-  typedef typename Container<MatrixType>::vector_type vector_type;
-  typedef typename Container<MatrixType>::partitioner_type partitioner_type;
-  typedef typename Container<MatrixType>::import_type import_type;
-
-  typedef typename Container<MatrixType>::HostView HostView;
-  typedef typename local_mv_type::dual_view_type::t_host HostViewLocal;
-
-  static_assert(std::is_same<MatrixType,
-                  Tpetra::RowMatrix<scalar_type, local_ordinal_type, global_ordinal_type, node_type> >::value,
-                "Ifpack2::BandedContainer: Please use MatrixType = Tpetra::RowMatrix.");
-
-  /// \brief The (base class) type of the input matrix.
-  ///
-  /// The input matrix to the constructor may be either a
-  /// Tpetra::RowMatrix specialization or a Tpetra::CrsMatrix
-  /// specialization.  However, we want to make the constructor as
-  /// general as possible, so we always accept the matrix as a
-  /// Tpetra::RowMatrix.  This typedef is the appropriate
-  /// specialization of Tpetra::RowMatrix.
-  typedef typename Container<MatrixType>::row_matrix_type row_matrix_type;
-
-  //@}
-public:
-  //! \name Constructor and destructor
-  //@{
-
-  /// \brief Constructor.
-  ///
-  /// \brief matrix [in] The original input matrix.  This Container
-  ///   will construct a local diagonal block from the rows given by
-  ///   <tt>localRows</tt>.
-  ///
-  /// \param localRows [in] The set of (local) rows assigned to this
-  ///   container.  <tt>localRows[i] == j</tt>, where i (from 0 to
-  ///   <tt>getNumRows() - 1</tt>) indicates the SparseContainer's
-  ///   row, and j indicates the local row in the calling process.
-  ///   <tt>localRows.size()</tt> gives the number of rows in the
-  ///   local matrix on each process.  This may be different on
-  ///   different processes.
-  ///   <tt>number of subdiagonals
-  ///   <tt>number of superdiagonals. Note: Internally, we store a Teuchos::SerialBandedMatrix
-  ///       with kl+ku superdiagonals, as we need the addtional storage for the LU decomposition.
-  BandedContainer (const Teuchos::RCP<const row_matrix_type>& matrix,
-                   const Teuchos::Array<Teuchos::Array<local_ordinal_type> >& partitions,
-                   const Teuchos::RCP<const import_type>& importer,
-                   int OverlapLevel,
-                   scalar_type DampingFactor);
-
-  BandedContainer (const Teuchos::RCP<const row_matrix_type>& matrix,
-                   const Teuchos::Array<local_ordinal_type>& localRows);
-
-  //! Destructor (declared virtual for memory safety of derived classes).
-  virtual ~BandedContainer ();
-
-  //@}
-  //! \name Get and set methods
-  //@{
-
-  //! Whether the container has been successfully initialized.
-  virtual bool isInitialized () const {
-    return IsInitialized_;
-  }
-
-  //! Whether the container has been successfully computed.
-  virtual bool isComputed () const {
-    return IsComputed_;
-  }
-
-  //! Set all necessary parameters.
-  virtual void setParameters (const Teuchos::ParameterList& List);
-
-  //@}
-  //! \name Mathematical functions
-  //@{
-
-  //! Do all set-up operations that only require matrix structure.
-  virtual void initialize ();
-
-  //! Initialize and compute each block.
-  virtual void compute ();
-
-  void clearBlocks();
-
-  //! Compute <tt>Y := alpha * M^{-1} X + beta*Y</tt>.
-  virtual void
-  apply (HostView& X,
-         HostView& Y,
-         int blockIndex,
-         int stride,
-         Teuchos::ETransp mode = Teuchos::NO_TRANS,
-         scalar_type alpha = Teuchos::ScalarTraits<scalar_type>::one(),
-         scalar_type beta = Teuchos::ScalarTraits<scalar_type>::zero()) const;
-
-  //! Compute <tt>Y := alpha * diag(D) * M^{-1} (diag(D) * X) + beta*Y</tt>.
-  virtual void
-  weightedApply (HostView& X,
-                 HostView& Y,
-                 HostView& D,
-                 int blockIndex,
-                 int stride,
-                 Teuchos::ETransp mode = Teuchos::NO_TRANS,
-                 scalar_type alpha = Teuchos::ScalarTraits<scalar_type>::one(),
-                 scalar_type beta = Teuchos::ScalarTraits<scalar_type>::zero()) const;
-
-  //@}
-  //! \name Miscellaneous methods
-  //@{
-
-  /// \brief Print information about this object to the given output stream.
-  ///
-  /// operator<< uses this method.
-  virtual std::ostream& print (std::ostream& os) const;
-
-  //@}
-  //! @name Implementation of Teuchos::Describable
-  //@{
-
-  //! A one-line description of this object.
-  virtual std::string description () const;
-
-  //! Print the object with some verbosity level to the given FancyOStream.
-  virtual void
-  describe (Teuchos::FancyOStream &out,
-            const Teuchos::EVerbosityLevel verbLevel =
-            Teuchos::Describable::verbLevel_default) const;
-
-  //@}
-
-  /// \brief Get the name of this container type for Details::constructContainer()
-  static std::string getName();
-
-private:
-  //! Copy constructor: Declared but not implemented, to forbid copy construction.
-  BandedContainer (const BandedContainer<MatrixType, LocalScalarType>& rhs);
-
-  //! Extract the submatrix identified by the local indices set by the constructor.
-  void extract ();
-
-  /// \brief Factor the extracted submatrix.
-  ///
-  /// Call this after calling extract().
-  void factor ();
-
-  /// \brief Post-permutation, post-view version of apply().
-  ///
-  /// apply() first does any necessary subset permutation and view
-  /// creation (or copying data), then calls this method to solve the
-  /// linear system with the diagonal block.
-  ///
-  /// \param X [in] Subset permutation of the input X of apply().
-  /// \param Y [in] Subset permutation of the input/output Y of apply().
-  void
-  applyImpl (HostViewLocal& X,
-             HostViewLocal& Y,
-             int blockIndex,
-             int stride,
-             Teuchos::ETransp mode,
-             const local_scalar_type alpha,
-             const local_scalar_type beta) const;
-
-  //! The local diagonal block, which compute() extracts.
-  std::vector<Teuchos::SerialBandDenseMatrix<int, local_scalar_type> > diagBlocks_;
-
-  //! Temporary X vector used in apply().
-  mutable std::vector<HostViewLocal> X_local;
-
-  //! Temporary Y vector used in apply().
-  mutable std::vector<HostViewLocal> Y_local;
-
-  //! Permutation array from LAPACK (GETRF).
-  Teuchos::Array<int> ipiv_;
-
-  //! If \c true, the container has been successfully initialized.
-  bool IsInitialized_;
-
-  //! If \c true, the container has been successfully computed.
-  bool IsComputed_;
-
-  Teuchos::Array<local_ordinal_type> kl_; //< number of subdiagonals
-  Teuchos::Array<local_ordinal_type> ku_; //< number of superdiagonals
-
-  //! Scalar data for all blocks
-  local_scalar_type* scalars_;
-
-  //! Offsets in scalars_ array for all blocks
-  Teuchos::Array<local_ordinal_type> scalarOffsets_;
-};
-
-}// namespace Ifpack2
+} // namespace Ifpack2
 
 #endif // IFPACK2_BANDEDCONTAINER_DECL_HPP