--- a/packages/ifpack2/src/Ifpack2_BandedContainer_decl.hpp
+++ b/packages/ifpack2/src/Ifpack2_BandedContainer_decl.hpp
@@ -46,7 +46,6 @@
 /// \brief Ifpack2::BandedContainer class declaration
 
 #include "Ifpack2_Container.hpp"
-#include "Ifpack2_Details_LapackSupportsScalar.hpp"
 #include "Tpetra_MultiVector.hpp"
 #include "Tpetra_Map.hpp"
 #include "Tpetra_RowMatrix.hpp"