Add DOF decoupling to MueLu/Ifpack2 line smoothing, improve Ifpack2 BlockRelaxation performance, and add BlockCrsMatrix support to all Ifpack2 container types

<!---
Be sure to select `develop` as the `base` branch against which to create this
pull request.  Only pull requests against `develop` will undergo Trilinos'
automated testing.  Pull requests against `master` will be ignored.

Provide a general summary of your changes in the Title above.  If this pull
request pertains to a particular package in Trilinos, it's worthwhile to start
the title with "PackageName:  ".

Note that anything between these delimiters is a comment that will not appear
in the pull request description once created. Most areas in this message are
commented out and can be easily added by removing the comment delimiters.

Please make sure to mark:
* Reviewers
* Assignees
* Labels

Replace <teamName> below with the appropriate Trilinos package/team name.
-->
@trilinos/ifpack2 

## Description

* adds block decoupling to Ifpack2. Set "block relaxation: decouple dofs" to true (default is false for backwards compatibility)
* lets TriDi, Banded and Sparse containers fully support Tpetra::BlockCrsMatrix as input
  (before, this was not supported but there was also no compile or runtime check that the matrix was not a BlockCrs)
* reduces Ifpack2 block relaxation setup time by 50% and
  apply time by 40% in Albany land ice problem
* reduces overall Albany time by 7%
* removes about 2K LOC of duplicated code by moving commonalities into
  Ifpack2_Container.hpp from the Container subclasses
* makes TriDi, Banded and Sparse containers fully support BlockCrs matrices (before,
  only Dense had this).

## Motivation and Context
<!--- Why is this change required?  What problem does it solve? -->
The Albany team noticed that in some problems that used line smoothing, the MueLu/Tpetra stack was up to 2x slower than ML/Epetra for the linear solves. In ML line smoothing, geometric lines are "decoupled" into one block per DOF, so that each block has the same sparsity pattern as the graph (tridiagonal). Ifpack2 block relaxation didn't have this feature, so tridiagonal blocks couldn't be used for any problems with > 1 dof/node. This PR adds optional decoupling to Ifpack2, and brings linear solve performance much closer to the ML stack.
<!---
If applicable, let us know how this merge request is related to any other open
issues or pull requests:

## Related Issues

* Closes 
* Blocks 
* Is blocked by 
* Follows 
* Precedes The Albany team noticed that in some land ice problems that used line smoothing, the MueLu/Tpetra stack was up to 2x slower than ML/Epetra for the linear solves. In ML line smoothing, geometric lines are "decoupled" into one block per DOF, so that each block has the same sparsity pattern as the graph (tridiagonal). Ifpack2 block relaxation didn't have this feature, so tridiagonal blocks couldn't be used for any problems with > 1 dof/node. This PR adds optional decoupling to Ifpack2, and makes linear solve performance compare much better against the ML stack.
* Related to 
* Part of 
* Composed of 
-->

## How Has This Been Tested?
<!---
Please describe in detail how you tested your changes.  Include details of your
testing environment and the tests you ran to see how your change affects other
areas of the code.  Consider including configure, build, and test log files.
-->
I added BlockRelaxation unit tests so that every combination of container (TriDi, Banded, Dense, Sparse), decoupling (on or off) and matrix type (Crs or BlockCrs) is tested.

## Albany land ice performance impact

Example performance data for Antartica ice sheet problem in Albany:

1 km resolution x 40 layers, 1.1B row linear systems. Run on 128 nodes/2048 cores of skybridge.

|                   | ML   | MueLu (PR) | MueLu (Current) | PR MueLu Speedup |
|-------------------|------|------------|-----------------|------------------|
| Albany Total Time | 2262 | 1920       | 2169            | 11.5%            |
| Linear Solve Time | 539  | 446        | 544             | 18%              |

8 km resolution x 5 layers, 2.5M row linear systems. Run on 1 node/16 cores of skybridge.

|                   | ML  | MueLu (PR) | MueLu (Current) | PR MueLu Speedup |
|-------------------|-----|------------|-----------------|------------------|
| Albany Total Time | 362 | 348        | 366             | 5%               |
| Linear Solve Time | 55  | 72         | 87              | 16.6%            |

<!---
Go over all the following points, and put an `x` in all the boxes that apply.
If you are unsure about any of these, please ask&mdash;we are here to help.
-->

## Checklist

- [ ] My commit messages mention the appropriate GitHub issue numbers.
- [x] My code follows the code style of the affected package(s).
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [ ] I have read the [code contribution guidelines](../blob/master/CONTRIBUTING.md) for this project.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] No new compiler warnings were introduced.
- [ ] These changes break backwards compatibility.

<!--- 
## Additional Information
Anything else we need to know in evaluating this merge request?
 -->